Add '2nd try' of the multiline parser:
Every extended response causes a flush of the previous mlbuf and starts
collecting new response lines.  Also, final_cb is now really only reached for
final responses and the code for case 'A' won't overwrite the command buffer
with the response anymore. I throw in '\n' as a separator for multi-line
responses, and the callback is called for each response this way.

(Philipp Zabel)