From: Andrzej Zaborowski <balrog@zabor.org> Date: Tue, 31 Jul 2007 22:28:28 +0200 Subject: [PATCH] Correctly split long usock reads into packets.
This is a patch for an unrelated to the above discussion timing issue in
libgsmd. In particular, when the modem responds fast enough, gsmd generates
usock packets fast enough for libgsmd to read multiple packets in a single read
and concatenate them thus discarding all packets but first. This happens on
Neo1973 when requesting for example the list of preferred operators or all
operators or, sometimes, list of short messages.