From: Andrzej Zaborowski <balrog@zabor.org> Date: Tue, 31 Jul 2007 22:26:36 +0200 Subject: [PATCH] Operations on the preferred operators list.
This makes use of the AT+CPOL variants to add / delete / list entries in the
preferred operators list on the SIM.  There's some inconsistency, similarly as
with operator selection, in that we return operator names in long
alphanumeric format but we take only operators in numeric format as
parameters.  Either the client will have to convert between the formats or
we should always use for example the long alphanumeric format, or maybe
do something else.