This patch adds a new MMC/SD MCI driver for the Glamo MMC/SD interface.
It supports

 - interrupt-based operation where possible
 - 1-bit and 4-bit mode
 - clock rate up to 16MHz -- 25MHz is not reliable and gives
     IO errors
 - read and write throughput is ~1.2MBytes/sec
 - voltage scaling as requested by MCI (done via platform)

The driver detects if it is on < GTA02 A5 and spins instead of using
the interrupt, so it will work on unmodified older A4 devices at least.

Several patches from this work fixing bugs in glamo-core and elsewhere
are already merged in mokopatches.  I have updated the kernel.git tree
as well with current mokopatches, so the only dependency at the moment
is glamo-platform-mmc.patch which is also in git.  You also need to
have these guys in .config:

CONFIG_MMC=y
CONFIG_MMC_BLOCK=y
CONFIG_MFD_GLAMO_MCI=y

You will typically see an error like this during boot

[   23.000000] FAT: invalid media value (0x15)
[   23.000000] VFS: Can't find a valid FAT filesystem on dev mmcblk0.
[   23.010000] FAT: invalid media value (0x15)
[   23.010000] VFS: Can't find a valid FAT filesystem on dev mmcblk0.

but it's okay, some flash devices don't have a partition table and it
tries to mount first the actual device node, only afterwards it tries
the first partition (mmcblk0p1), which succeeds silently.

One remaining write corruption bug with Glamo MMC behaviour is not
certain to be solved yet, the following stress test will expose any
problems:

$ dd if=/dev/urandom of=/tmp/dump bs=4096 count=2560
2560+0 records in
2560+0 records out
$ mount /media/card ; c=0; hexdump -C /tmp/dump >/tmp/dump.txt ; a=`md5sum /tmp/dump | cut -d' ' -f1` ; while [ 1 ] ; do cp /tmp/dump /media/card ; umount /media/card ; mount /media/card ; b=`md5sum /media/card/dump | cut -d' ' -f1`; if [ "$a" != "$b" ] ; then echo $c: ***** FAIL $a != $b ; hexdump -C /media/card/dump >/tmp/mc-dump.txt; diff -urN /tmp/dump.txt /tmp/mc-dump.txt | head -n50; else echo $c: OK ; fi ; c=$(( $c + 1 )); done

It creates a 10MB random file and then copies it to the card, unmounts
forcing it to be flushed to the card, remounts and checks the md5sum
is the same as before.  If not, it shows the first part of an ascii hex
diff of the change.

Signed-off-by: Andy Green <andy@openmoko.com>

smedia-glamo.patch:
- drivers/mfd/glamo/glamo-core.c (glamo_init_script): more register changes 
glamo-mmc.patch:
- drivers/mfd/glamo/Kconfig, drivers/mfd/glamo/glamo-core.c, 
  drivers/mfd/glamo/glamo-mci.c, drivers/mfd/glamo/glamo-mci.h,
  drivers/mfd/glamo/glamo-regs.h: SD/MMC driver