Applied funbar-2.6.24-platform-spi-gpio-acc-repair.patch to gta02-acc.patch:
This doesn't actually do anything about the INT problem yet.
The 2.6.24 gpio spi and platform situation was completely
messed up and was a long way from even being able to talk
to the motion sensors.

This patch gets the gpio SPI actually working and both
motion sensors probed and recognized on 2.6.24.

(FUNBAR is like FUBAR with a "Nearly" in there.)

 - the gpio cs function only addressed one chip, I
   pushed the pin info through platform so it does
   both motion sensors

 - the spi_board_info chip_select member ain't a
   pin ID, it is a chip select index for the
   spi master

 - nobody set master->num_chipselect, I pushed it
   through platform stuff into s3c2410_spigpio_info

 - name of the gpio spi driver had changed and the
   platform side not updated

Signed-off-by: Andy Green <andy@openmoko.com>