If we have a writeable environment in persistent storage and CONFIG_CMD_LOADENV is set, add the command "loadenv" to load a new environment.
This is to be used for getting access to the persistent environment stored in
NAND after we've booted the GTA02 from NOR.

This patch also adds a command "defaultenv" that resets the environment to the
hard-coded default settings, which is also useful without "loadenv".

Note that we don't scan the new environment for changes in settings (e.g.,
"stdout"). This is to ensure that we don't accidently cut ourselves off during
environmental surgery.

Signed-off-by: Werner Almesberger <werner@openmoko.org>

loadenv.patch:
- common/cmd_nvedit.c (do_loadenv): load the environment from persistent
  storage (without updating settings derived from the environment)
- common/cmd_nvedit.c (do_defaultenv): reset the environment to default
  settings
- include/configs/neo1973_gta02.h: added CONFIG_CMD_LOADENV and
  CONFIG_CMD_DEFAULTENV

series: added loadenv.patch and also gta02-bootmenu.patch