ar6k: Beacon filtering enabled once scan is done
When we're done with scanning, the host doesn't need to get all the
beacons anymore. Later on, we could fine tune that to only get our
BSS beacons (mandatory for roaming).

But for now, we just filter them all after scanning is done so that
we don't get dozens of interrupts per seconds because of that.

Signed-off-by: Samuel Ortiz <sameo@openedhand.com>

---
 drivers/sdio/function/wlan/ar6000/ar6000/wireless_ext.c |    5 +++++
 1 file changed, 5 insertions(+)