fix-pcf50633-LOWBAT-kill-init.patch
(Corrected a & vs. == precedence inversion.)

From: Andy Green <andy@openmoko.com>

Until now the driver treats LOWBAT as a fatal crisis and responds
by going postal on init.  But the driver didn't think that we can
have adapter or USB power and no battery, which is fine.

This patch reserves the crisis behaviour for when we have no
adapter / USB power and just acknowledges the PMU exception when
we do.

Signed-off-by: Andy Green <andy@openmoko.com>

---

 drivers/i2c/chips/pcf50633.c |   24 +++++++++++++++++-------
 1 files changed, 17 insertions(+), 7 deletions(-)