fix-pcf50633-one-shot-resume.patch
From: Andy Green <andy@openmoko.com>

The wakeup sources behind the PMU interrupt only worked once
as a wakeup source for suspend.  This is because of further
resume sequencing trouble meaning that the PMU interrupt
gets accepted for service before the I2C driver resumed and
it is able to actually talk to the PMU.  A previous patch
stopped this situation stalling resume, but this patch
extends it to scheduling a retry after an msleep(1).

Failing to service the PMU interrupt led to no further PMU
interrupts being recognized due to edge triggering, with the
retry we typically get it on the next try since I2C is up
by then, and PMU interrupts continue to be recognized for
wake and other functions.

Signed-off-by: Andy Green <andy@openmoko.com>

---

 drivers/i2c/chips/pcf50633.c |   22 ++++++++++++++++++++--
 1 files changed, 20 insertions(+), 2 deletions(-)