Subject: [PATCH] fix-glamo-reissue-memory-settings-during-resume.patch
The "black snow" effect after resume seems to be caused by the
memory in the glamo needing to be reset after the
"deep power down".  This patch rewrites a subset of registers in
that area on resume (reinitializing them all destroys the video
output action) and removes the black snow issue.

The "black snow" is logical corruption of the pixel data in the
glamo confirmed by locking a scope to VYSNC and looking at the
pixel data at the same offset over multiple frames, it was
changing frame by frame logically for the same offset.  So it
makes sense the underlying cause of problems here is the memory
in the Glamo.

Also, sometimes when the SD card was kicked into 16MHz speed after
resume it failed, this is also believed to be to the Glamo internal
memory which the SD peripheral also uses being in a bad state on
resume and this might also be helped.

Signed-off-by: Andy Green <andy@openmoko.com>
---

 drivers/mfd/glamo/glamo-core.c |   93 ++++++++++++++++++++++++++++++----------
 drivers/mfd/glamo/glamo-core.h |    2 -
 2 files changed, 71 insertions(+), 24 deletions(-)